Calibrates SOFUN model parameters (for parallelized applications)

Runs the requested model calibration and returns results

calib_sofun(
  drivers,
  obs,
  settings_calib,
  optim_out = TRUE,
  logpath = file.path(tempdir(), paste0("out_calib_", "my_calibration_name",
    ".rds.log.txt")),
  ...
)

Arguments

drivers

A data frame with driver data. See pmodel_drivers for a description of the data structure. Additional columns can optionally be provided to drivers to control e.g. the processing within a personalized cost function.

obs

A data frame containing observational data used for model calibration. See pmodel_validation for a description of the data structure. Additional columns can optionally be provided to obs to control e.g. the processing within a personalized cost function.

settings_calib

A list containing model calibration settings. See the 'P-model usage' vignette for more information and examples.

method

A string indicating the optimization method, either 'GenSA' or 'BayesianTools'.

par

A list of model parameters. For each parameter, an initial value and lower and upper bounds should be provided. The calibratable parameters include model parameters 'kphio', 'kphio_par_a', 'kphio_par_b', 'soilm_thetastar', 'soilm_betao', 'beta_costunitratio', 'rd_to_vcmax', 'tau_acclim', 'kc_jmax' and 'rootzone_whc' , and (if doing Bayesian calibration) error parameters for each target variable, named for example 'err_gpp'. This list must match the input parameters of the calibration metric and the parameters should be given in the order above.

metric

A cost function. See the 'Cost functions for parameter calibration' vignette for examples.

control

A list of arguments passed on to the optimization function. If method = 'GenSA', see GenSA. If method = 'BayesianTools' the list should include at least sampler_runMCMC and settings_runMCMC, see BayesianTools::runMCMC. For method = 'BayesianTools', n_parallel_nrChains can optionally be used to specify the number of cores to be used for parallel sampling of chains. (If >=2 it overwrites settings_runMCMC$nrChains., if 0 or 1 parallel sampling is deactivated.

optim_out

A logical indicating whether the function returns and stores the raw output of the optimization functions (defaults to TRUE).

logpath

A path where a log-file is created and progress logged for parallel sampling, defaults to a temporary directory. The path is stored in the output object as simulation `name` and `logpath`. (Note that this argument is only used when parallelized sampling is requested.)

...

Optional arguments, simply passed on to the cost function.

Value

A named list containing the calibrated parameter vector `par`, the output object from the optimization `mod`, and additional identifier outputs `name`, `logpath`, and `walltime`. For more details on the `mod` output and how to evaluate it, see runMCMC (also this post) and GenSA. If parallel (`n_parallel_nrChains >= 2`) also a logfile is created at `logpath`.

Details

This is the main function that handles the calibration of SOFUN model parameters.

Examples

# Calib 1: use Bayesian calibration approach with likelihood:
# Define priors of model parameters that will be calibrated
params_to_estimate <- list(
  kphio           = list(lower = 0.02, upper = 0.15, init = 0.05),
  err_gpp         = list(lower = 0.01, upper = 3, init = 0.8),
  err_bigD13C     = list(lower = 0.5, upper = 4.0, init = 2.0)
)
# Fix model parameters that won't be calibrated
params_fix       <- list(
  kphio_par_a        = 0,
  kphio_par_b        = 25,
  soilm_thetastar    = 0.6 * 240,
  soilm_betao        = 0.01,
  beta_unitcostratio = 146,
  rd_to_vcmax        = 0.014,
  tau_acclim         = 30,
  kc_jmax            = 0.41
)
# Define calibration settings
settings_calib <- list(
  method  = "BayesianTools",
  par     = params_to_estimate,
  metric  = rsofun::cost_likelihood_pmodel,
  control = list(
    sampler_runMCMC = "DEzs",
    settings_runMCMC = list(
      burnin = 0,
      iterations = 10,     # kept artificially low
      nrChains = 1         # 2, # these are the independent chains
    ),
    n_parallel_nrChains = 1  # 2, this can be parallelized
  )
)
# Run the calibration for GPP and D13C data
drivers_to_use <- rsofun::pmodel_drivers    |>
  dplyr::filter(sitename %in% c("FR-Pue", "lon_+146.13_lat_-032.97"))
obs_to_use     <- rsofun::pmodel_validation |>
  dplyr::filter(sitename %in% c("FR-Pue", "lon_+146.13_lat_-032.97"))
calib_output <- rsofun::calib_sofun(
  drivers = drivers_to_use,
  obs     = obs_to_use,
  settings_calib = settings_calib,
  # extra arguments for the cost function
  par_fixed = params_fix
)
#>  Running DEzs-MCMC, chain  1 iteration 12 of 12 . Current logp  -11241.39 -4028.086 -3639.182 . Please wait! 
#> runMCMC terminated after 0.667999999999999seconds
calib_output$mod      # BayesianTools::mcmcSamplerList
#> [1] "mcmcSamplerList - you can use the following methods to summarize, plot or reduce this class:"
#> [1] plot    print   summary
#> see '?methods' for accessing help and source code
calib_output$par      # Named vector
#>       kphio     err_gpp err_bigD13C 
#>  0.04862024  1.32409502  2.55882781 
calib_output$walltime # Benchmarked time
#> Time difference of 0.6701379 secs
calib_output$runtime  # unused
#> NULL
calib_output$name     # optionally used calibration name
#> [1] "out_calib_my_calibration_name.rds.log.txt"
calib_output$fpath    # path of rds output
#> NULL

# Calib 2: use GenSa optimization of RMSE
# Calibrate the model and optimize the free parameters using demo datasets
settings_calib_rmse <- list(
  method = "GenSA",                   # minimizes the RMSE
  metric = cost_rmse_pmodel,          # our cost function returning the RMSE
  control = list( # control parameters for optimizer GenSA
    maxit = 2),
  par = list( # bounds for the parameter space
    kphio = list(lower = 0.02, upper = 0.2, init = 0.05)
  )
)
drivers_to_use <- pmodel_drivers
obs_to_use     <- pmodel_validation
pars_calib_rmse <- calib_sofun(
  # calib_sofun arguments:
  drivers  = drivers_to_use,
  obs      = obs_to_use,
  settings_calib = settings_calib_rmse,
  # extra arguments passed to the cost function:
  par_fixed = list( # fix all other parameters
    kphio_par_a        = 0.0,        # set to zero to disable temperature-dependence
    # of kphio, setup ORG
    kphio_par_b        = 1.0,
    soilm_thetastar    = 0.6 * 240,  # to recover paper setup with soil moisture stress
    soilm_betao        = 0.0,
    beta_unitcostratio = 146.0,
    rd_to_vcmax        = 0.014,      # value from Atkin et al. 2015 for C3 herbaceous
    tau_acclim         = 30.0,
    kc_jmax            = 0.41
  )
)