Creates a log-likelihood cost function for biomee with different targets

Creates a cost function for parameter calibration, which computes the log-likelihood for the biomee model fitting several target variables for a given set of parameters.

create_cost_likelihood_biomee(targets)

Arguments

targets

A character vector indicating the target variables for which the optimization will be done. This should be a subset of c("GPP", "LAI", "Density", "Biomass").

Value

A cost function which computes the log-likelihood of the simulated targets by the biomee model versus the observed targets. This cost function has as arguments a vector of calibratable model parameters par, a data frame of observations obs, and a data frame of driver data drivers.

Details

The resulting cost function performs a biomee model run for the value of par given as argument. par must always contain 'phiRL', 'LAI_light', 'tf_base', 'par_mort' and the error terms corresponding to the target variables, e.g. 'err_GPP' if GPP is a target. Make sure that the order of the error terms in par coincides with the order of the targets used to create the cost function.

The likelihood is calculated assuming that the predicted targets are independent, normaly distributed and centered on the observations. All targets have the same weight in the optimization objective. The optimization should be run using BayesianTools, and the likelihood is maximized.