Creates a cost function for different simulation setups based on RMSE

Creates a cost function for parameter calibration, keeping non-calibrated parameter values fixed and calibrating the parameters corresponding to setups BRC and FULL from Stocker et al., 2020 GMD. The cost function computes the root mean squared error (RMSE) on the calibrated parameters.

create_cost_rmse_pmodel(params_modl, setup, method)

Arguments

params_modl

A list of model parameter values, including 'kphio', 'soilm_par_a', 'soilm_par_b', 'tau_acclim_tempstress' and 'par_shape_tempstress' in that order.

setup

A character string ('BRC' or 'FULL') indicating which parameters are calibrated. For setup = 'BRC' only the quantum yield efficiency kphio is calibrated; for setup = 'FULL' it also includes the soil moisture stress parameters soilm_par_a and soilm_par_b for calibration.

method

A character string indicating the optimization method that will be used, either 'BayesianTools' or 'GenSA'.

Value

A cost function which computes the RMSE of the simulated GPP by the P-model versus the observed GPP. This cost function has as arguments a list of calibratable model parameters par, a data frame of observations obs, and a data frame of driver data drivers.

Details

The resulting cost function performs a P-model run for the value of par given as argument and the remaining non-calibratable parameters are held constant (specified via params_modl).

Since the calibration routine in BayesianTools is based on maximizing a cost function and we want to minimize the RMSE, the opposite value, (-1)*RMSE, is returned if method = 'BayesianTools'. GenSA minimizes the given objective function, so the plain RMSE is returned when method = 'GenSA'.

Examples

if (FALSE) {
# Set model parameters
pars <- list(
  kphio          = 0.04,
  soilm_par_a    = 2.8,
  soilm_par_b    = 1.7,
  tau_acclim_tempstress  = 7.3,
  par_shape_tempstress   = 0.1
  )

# Write cost function
cost_rmse_kphio <- create_cost_rmse_pmodel(
  params_modl = pars,
  setup = 'BRC',
  method = 'BayesianTools'
  )
}