rsofun P-model GPP and leaf D13C validation data

Small example dataset of target observations (daily GPP flux data as well as leaf D13C data) to optimize model parameters with the function calib_sofun

pmodel_validation

Format

A tibble of validation data:

sitename

A character string containing the site name (e.g. 'FR-Pue').

targets

A single one-row tibble with a column for each target containing TRUE or FALSE (or NA_logical)

data

A single tibble or a list of tibbles with target observations. For daily simulations this single tibble can contain daily values for:

date

Date vector with format YYYY-MM-DD.

gpp

The observed Gross Primary Productivity (GPP) for each time stamp (in gC m\(^{-2}\) d\(^{-1}\)).

gpp_qc

A quality check indicator.

le

Latent heat of exchange ...

...

... etc.

For onestep acclimation this can fore example include:

bigD13C

A tibble containing observations of leaf D13C values.

Vcmax

A tibble containing observations of Vcmax values.

Jmax

A tibble containing observations of Jmax values.

Source

Pastorello, G., Trotta, C., Canfora, E. et al. The FLUXNET2015 dataset and the ONEFlux processing pipeline for eddy covariance data. Sci Data 7, 225 (2020). https://doi.org/10.1038/s41597-020-0534-3

Cornwell, W. (2025). traitecoevo/leaf13C: V0.2.2 (Version v0.2.2) [Computer software]. Zenodo. https://doi.org/10.5281/zenodo.15239220

Cornwell, William K., et al. “Climate and soils together regulate photosynthetic carbon isotope discrimination within C3 plants worldwide.” Global Ecology and Biogeography 27.9 (2018): 1056-1067.

Examples

require(ggplot2)
#> Loading required package: ggplot2
require(tidyr)
#> Loading required package: tidyr
p_model_oldformat_validation %>% tidyr::unnest(data)
#> # A tibble: 2,190 × 4
#>    sitename date         gpp gpp_unc
#>    <chr>    <date>     <dbl>   <dbl>
#>  1 FR-Pue   2007-01-01 2.21  0.0108 
#>  2 FR-Pue   2007-01-02 2.23  0.00475
#>  3 FR-Pue   2007-01-03 2.48  0.00727
#>  4 FR-Pue   2007-01-04 1.71  0.00516
#>  5 FR-Pue   2007-01-05 2.61  0.00764
#>  6 FR-Pue   2007-01-06 2.85  0.00610
#>  7 FR-Pue   2007-01-07 2.80  0.0164 
#>  8 FR-Pue   2007-01-08 1.90  0.0116 
#>  9 FR-Pue   2007-01-09 3.42  0.0147 
#> 10 FR-Pue   2007-01-10 0.651 0.0151 
#> # ℹ 2,180 more rows
pmodel_validation |> tidyr::unnest_wider(targets, names_sep = "_")
#> # A tibble: 20 × 4
#>    sitename                targets_1 targets_2 data                
#>    <chr>                   <chr>     <chr>     <list>              
#>  1 CH-Dav                  gpp       le        <tibble [714 × 7]>  
#>  2 CZ-BK1                  gpp       NA        <tibble [479 × 7]>  
#>  3 DK-Sor                  gpp       NA        <tibble [720 × 7]>  
#>  4 FI-Hyy                  gpp       NA        <tibble [704 × 7]>  
#>  5 FR-Pue                  gpp       NA        <tibble [1,924 × 7]>
#>  6 GF-Guy                  gpp       NA        <tibble [319 × 7]>  
#>  7 US-Ha1                  gpp       NA        <tibble [502 × 7]>  
#>  8 US-PFa                  gpp       NA        <tibble [669 × 7]>  
#>  9 lon_+010.52_lat_+051.08 bigD13C   NA        <tibble [8 × 2]>    
#> 10 lon_+011.10_lat_+048.30 bigD13C   NA        <tibble [1 × 2]>    
#> 11 lon_+112.58_lat_+023.13 bigD13C   NA        <tibble [94 × 2]>   
#> 12 lon_+145.13_lat_-005.83 bigD13C   NA        <tibble [1 × 2]>    
#> 13 lon_+146.13_lat_-032.97 bigD13C   NA        <tibble [19 × 2]>   
#> 14 lon_+148.30_lat_-036.10 bigD13C   NA        <tibble [2 × 2]>    
#> 15 lon_+153.00_lat_-026.85 bigD13C   NA        <tibble [25 × 2]>   
#> 16 lon_-079.10_lat_+035.97 bigD13C   NA        <tibble [4 × 2]>    
#> 17 lon_-116.45_lat_+047.16 bigD13C   NA        <tibble [18 × 2]>   
#> 18 lon_-119.82_lat_+034.50 bigD13C   NA        <tibble [13 × 2]>   
#> 19 lon_-122.98_lat_+038.40 bigD13C   NA        <tibble [1 × 2]>    
#> 20 lon_-149.61_lat_+063.97 bigD13C   NA        <tibble [1 × 2]>    
pmodel_validation |> dplyr::filter(grepl("^[A-Z]", sitename)) |> unnest(data)
#> # A tibble: 6,031 × 9
#>    sitename targets   date          gpp gpp_qc   nee nee_qc    le le_qc
#>    <chr>    <list>    <date>      <dbl>  <dbl> <dbl>  <dbl> <dbl> <dbl>
#>  1 CH-Dav   <chr [2]> 2007-01-01  0.855  1     1.44   1      9.49     1
#>  2 CH-Dav   <chr [2]> 2007-01-02 -0.943  1     2.62   1     65.0      1
#>  3 CH-Dav   <chr [2]> 2007-01-03 -0.951  1     2.39   1     48.6      1
#>  4 CH-Dav   <chr [2]> 2007-01-04  1.23   0.979 0.958  0.979 14.7      1
#>  5 CH-Dav   <chr [2]> 2007-01-05  1.36   1     0.992  1     37.5      1
#>  6 CH-Dav   <chr [2]> 2007-01-06  1.67   1     1.02   1     31.4      1
#>  7 CH-Dav   <chr [2]> 2007-01-07  2.27   0.979 0.575  0.979 16.9      1
#>  8 CH-Dav   <chr [2]> 2007-01-08  2.00   1     0.854  1     21.5      1
#>  9 CH-Dav   <chr [2]> 2007-01-09  1.20   0.979 1.53   0.979 24.6      1
#> 10 CH-Dav   <chr [2]> 2007-01-10  0.732  0.958 1.72   0.958 28.3      1
#> # ℹ 6,021 more rows
pmodel_validation |> dplyr::filter(grepl("lon", sitename)) |> unnest(data)
#> # A tibble: 187 × 4
#>    sitename                targets   id                       bigD13C
#>    <chr>                   <list>    <chr>                      <dbl>
#>  1 lon_+010.52_lat_+051.08 <chr [1]> 2001_Acer platanoides       19.4
#>  2 lon_+010.52_lat_+051.08 <chr [1]> 2001_Acer pseudoplatanus    19.7
#>  3 lon_+010.52_lat_+051.08 <chr [1]> 2001_Carpinus betulus       19.4
#>  4 lon_+010.52_lat_+051.08 <chr [1]> 2001_Fagus engleriana       19.5
#>  5 lon_+010.52_lat_+051.08 <chr [1]> 2001_Fraxinus excelsior     18.0
#>  6 lon_+010.52_lat_+051.08 <chr [1]> 2001_Quercus x mannifera    20.0
#>  7 lon_+010.52_lat_+051.08 <chr [1]> 2001_Tilia cordata          19.4
#>  8 lon_+010.52_lat_+051.08 <chr [1]> 2001_Tilia x europaea       18.2
#>  9 lon_+011.10_lat_+048.30 <chr [1]> 2002_Fagus engleriana       22.2
#> 10 lon_+112.58_lat_+023.13 <chr [1]> 1985_Adenosma glutinosa     22.8
#> # ℹ 177 more rows