R wrapper for SOFUN biomee

Call to the biomee Fortran model

run_biomee_f_bysite(
  sitename,
  params_siml,
  site_info,
  forcing,
  params_tile,
  params_species,
  params_soil,
  init_cohort,
  init_soil,
  makecheck = TRUE
)

Arguments

sitename

Site name.

params_siml

Simulation parameters.

spinup

A logical value indicating whether this simulation does spin-up.

spinupyears

Number of spin-up years.

recycle

Length of standard recycling period (days).

firstyeartrend

First transient year.

nyeartrend

Number of transient years.

outputhourly

A logical value indicating whether hourly output is produced.

outputdaily

A logical value indicating whether daily output is produced.

do_U_shaped_mortality

A logical value indicating whether U-shaped mortality is used.

update_annualLAImax

A logical value indicating whether updating LAImax according to mineral N in soil.

do_closedN_run

A logical value indicating whether doing N closed runs to recover N balance.

do_reset_veg

A logical value indicating whether reseting vegetation for disturbance runs.

dist_frequency

Value indicating the frequency of the disturbance event (in years) (e.g. 100 indicates a disturbance event occurs every 100 years, i.e., at a rate of 0.01)

code_method_photosynth

String specifying the method of photosynthesis used in the model, either "pmodel" or "gs_leuning".

code_method_mortality

String indicating the type of mortality in the model. One of the following: "dbh" is size-dependent mortality, "const_selfthin" is constant self thinning (in development), "cstarvation" is carbon starvation, and "growthrate" is growth rate dependent mortality.

site_info

Site meta info in a data.frame.

sitename

Name of the site.

lon

Longitud of the site location.

lat

Latitude of the site location.

elv

Elevation of the site location, in meters.

forcing

Forcing data.frame used as input.

params_tile

Tile-level model parameters, into a single row data.frame with columns:

soiltype

Integer indicating the type of soil: Sand = 1, LoamySand = 2, SandyLoam = 3, SiltLoam = 4, FrittedClay = 5, Loam = 6, Clay = 7.

FLDCAP

Field capacity (vol/vol). Water remaining in a soil after it has been thoroughly saturated and allowed to drain freely.

WILTPT

Wilting point (vol/vol). Water content of a soil at which plants wilt and fail to recover.

K1

Fast soil C decomposition rate (year\(^{-1}\)).

K2

Slow soil C decomposition rate (year\(^{-1}\)).

K_nitrogen

Mineral Nitrogen turnover rate (year\(^{-1}\)).

MLmixRatio

Ratio of C and N returned to litters from microbes.

etaN

N loss rate through runoff (organic and mineral) (year\(^{-1}\)).

LMAmin

Minimum LMA, leaf mass per unit area, kg C m\(^{-2}\).

fsc_fine

Fraction of fast turnover carbon in fine biomass.

fsc_wood

Fraction of fast turnover carbon in wood biomass.

GR_factor

Growth respiration factor.

l_fract

Fraction of the carbon retained after leaf drop.

retransN

Retranslocation coefficient of nitrogen.

f_initialBSW

Coefficient for setting up initial sapwood.

f_N_add

Re-fill of N for sapwood.

tf_base

Calibratable scalar for respiration, used to increase LUE levels.

par_mort

Canopy mortality parameter.

par_mort_under

Parameter for understory mortality.

params_species

A data.frame containing species-specific model parameters, with one species per row. The columns of this data.frame are:

lifeform

Integer set to 0 for grasses and 1 for trees.

phenotype

Integer set to 0 for deciduous and 1 for evergreen.

pt

Integer indicating the type of plant according to photosynthesis: 0 for C3; 1 for C4

alpha_FR

Fine root turnonver rate (year\(^{-1}\)).

rho_FR

Material density of fine roots (kg C m\(^{-3}\)).

root_r

Radious of the fine roots, in m.

root_zeta

e-folding parameter of root vertical distribution, in m.

Kw_root

Fine root water conductivity (mol m\(^{-2}\) s\(^{-1}\) MPa\(^{-1}\)).

leaf_size

Characteristic leaf size.

Vmax

Max RuBisCo rate, in mol m\(^{-2}\) s\(^{-1}\).

Vannual

Annual productivity per unit area at full sun (kg C m\(^{-2}\) year\(^{-2}\)).

wet_leaf_dreg

Wet leaf photosynthesis down-regulation.

m_cond

Factor of stomatal conductance.

alpha_phot

Photosynthesis efficiency.

gamma_L

Leaf respiration coefficient, in year\(^{-1}\).

gamma_LN

Leaf respiration coefficient per unit N.

gamma_SW

Sapwood respiration rate, in kg C m\(^{-2}\) year\(^{-1}\).

gamma_FR

Fine root respiration rate, kg C kg C\(^{-1}\) year\(^{-1}\).

tc_crit

Critical temperature triggerng offset of phenology, in Kelvin.

tc_crit_on

Critical temperature triggerng onset of phenology, in Kelvin.

gdd_crit

Critical value of GDD5 for turning ON growth season.

betaON

Critical soil moisture for phenology onset.

betaOFF

Critical soil moisture for phenology offset.

seedlingsize

Initial size of seedlings, in kg C per individual.

LNbase

Basal leaf N per unit area, in kg N m\(^{-2}\).

lAImax

Maximum crown LAI (leaf area index).

Nfixrate0

Reference N fixation rate (kg N kg C\(^{-1}\) root).

NfixCost0

Carbon cost of N fixation (kg C kg N\(^{-1}\)).

phiCSA

Ratio of sapwood area to leaf area.

mortrate_d_c

Canopy tree mortality rate (year\(^{-1}\)).

mortrate_d_u

Understory tree mortality rate (year\(^{-1}\)).

maturalage

Age at which trees can reproduce (years).

v_seed

Fraction of G_SF to G_F.

fNSmax

Multiplier for NSNmax as sum of potential bl and br.

LMA

Leaf mass per unit area (kg C m\(^{-2}\)).

rho_wood

Wood density (kg C m\(^{-3}\)).

alphaBM

Coefficient for allometry (biomass = alphaBM * DBH ** thetaBM).

thetaBM

Coefficient for allometry (biomass = alphaBM * DBH ** thetaBM).

kphio

Quantum yield efficiency \(\varphi_0\), in mol mol\(^{-1}\).

phiRL

Ratio of fine root to leaf area.

LAI_light

Maximum LAI limited by light.

params_soil

A tibble of soil parameters (one row per soil layer).

type

A string indicating the type of soil.

GMD

Geometric mean particle diameter (mm).

GSD

Geometric standard deviation of particle size.

vwc_sat

Saturated volumetric soil water content (vol/vol).

chb

Soil texture parameter.

psi_sat_ref

Saturation soil water potential (m).

k_sat_ref

Hydraulic conductivity of saturated soil (kg m\(^{-2}\) s\(^{-1}\)).

alphaSoil

Vertical changes of soil property, where 1 = no change.

heat_capacity_dry

Heat capacity dry air (J m\(^{-3}\) K\(^{-1}\)).

init_cohort

A data.frame of initial cohort specifications.

init_cohort_species

Indicates different species.

init_cohort_nindivs

Initial individual density, in individuals per m\(^{2}\).

init_cohort_bsw

Initial biomass of sapwood, in kg C per individual.

init_cohort_bHW

Initial biomass of heartwood, in kg C per tree.

init_cohort_nsc

Initial non-structural biomass.

init_soil

A data.frame of initial soil pools.

init_fast_soil_C

Initial fast soil carbon, in kg C m\(^{-2}\).

init_slow_soil_C

Initial slow soil carbon, in kg C m\(^{-2}\).

init_Nmineral

Mineral nitrogen pool, in kg N m\(^{-2}\).

N_input

Annual nitrogen input to soil N pool, in kg N m\(^{-2}\) year\(^{-1}\).

makecheck

A logical specifying whether checks are performed to verify forcings.

Value

Model output is provided as a list, with elements:

output_hourly_tile

A data.frame containing hourly predictions .

year

Year of the simulation.

doy

Day of the year.

hour

Hour of the day.

rad

Radiation, in W m\(^{-2}\).

Tair

Air temperature, in Kelvin.

Prcp

Precipitation, in mm m\(^{-2}\).

GPP

Gross primary production (kg C m\(^{-2}\) hour\(^{-1}\)).

Resp

Plant respiration (kg C m\(^{-2}\) hour\(^{-1}\)).

Transp

Transpiration (mm m\(^{-2}\)).

Evap

Evaporation (mm m\(^{-2}\)).

Runoff

Water runoff (mm m\(^{-2}\)).

Soilwater

Soil water content in root zone (kg m\(^{-2}\)).

wcl

Volumetric soil water content for each layer (vol/vol).

FLDCAP

Field capacity (vol/vol).

WILTPT

Wilting point (vol/vol).

output_daily_tile

A data.frame with daily outputs at a tile level.

year

Year of the simulation.

doy

Day of the year.

Tc

Air temperature (Kelvin).

Prcp

Precipitation (mm m\(^{-2}\)).

totWs

Soil water content in root zone (kg m\(^{-2}\)).

Trsp

Transpiration (mm m\(^{2-}\)).

Evap

Evaporation (mm m\(^{-2}\)).

Runoff

Water runoff (mm m\(^{-2}\)).

ws1

Volumetric soil water content for layer 1.

ws2

Volumetric soil water content for layer 2.

ws3

Volumetric soil water content for layer 3.

LAI

Leaf area index (m\(^2\)/m\(^2\)).

GPP

Gross primary production (kg C m\(^{-2}\) day\(^{-1}\)).

Rauto

Plant autotrophic respiration (kg C m\(^{-2}\) day\(^{-1}\)).

Rh

Heterotrophic respiration (kg C m\(^{-2}\) day\(^{-1}\)).

NSC

Non-structural carbon (kg C m\(^{-2}\)).

seedC

Biomass of seeds (kg C m\(^{-2}\)).

leafC

Biomass of leaves (kg C m\(^{-2}\)).

rootC

Biomass of fine roots (kg C m\(^{-2}\)).

SW_C

Biomass of sapwood (kg C m\(^{-2}\)).

HW_C

biomass of heartwood (kg C m\(^{-2}\)).

NSN

Non-structural N pool (kg N m\(^{-2}\)).

seedN

Nitrogen of seeds (kg N m\(^{-2}\)).

leafN

Nitrogen of leaves (kg N m\(^{-2}\)).

rootN

Nitrogen of roots (kg N m\(^{-2}\)).

SW_N

Nitrogen of sapwood (kg N m\(^{-2}\)).

HW_N

Nitrogen of heartwood (kg N m\(^{-2}\)).

McrbC

Microbial carbon (kg C m\(^{-2}\)).

fastSOM

Fast soil carbon pool (kg C m\(^{-2}\)).

slowSOM

Slow soil carbon pool (kg C m\(^{-2}\)).

McrbN

Microbial nitrogen (kg N m\(^{-2}\)).

fastSoilN

Fast soil nitrogen pool (kg N m\(^{-2}\)).

slowSoilN

Slow soil nitrogen pool (kg N m\(^{-2}\)).

mineralN

Mineral nitrogen pool (kg N m\(^{-2}\)).

N_uptk

Nitrogen uptake (kg N m\(^{-2}\)).

output_daily_cohorts

A data.frame with daily predictions for each canopy cohort.

year

Year of the simulation.

doy

Day of the year.

hour

Hour of the day.

cID

An integer indicating the cohort identity.

PFT

An integer indicating the Plant Functional Type.

layer

An integer indicating the crown layer, numbered from top to bottom.

density

Number of trees per area (trees ha\(^{-1}\)).

f_layer

Fraction of layer area occupied by this cohort.

LAI

Leaf area index (m\(^2\)/m\(^2\)).

gpp

Gross primary productivity (kg C tree\(^{-1}\) day\(^{-1}\)).

resp

Plant autotrophic respiration (kg C tree\(^{-1}\) day\(^{-1}\)).

transp

Transpiration (mm tree\(^{-1}\) day\(^{-1}\)).

NPPleaf

Carbon allocated to leaves (kg C tree\(^{-1}\) day\(^{-1}\)).

NPProot

Carbon allocated to fine roots (kg C tree\(^{-1}\) day\(^{-1}\)).

NPPwood

Carbon allocated to wood (kg C tree\(^{-1}\) day\(^{-1}\)).

NSC

Nonstructural carbohydrates of a tree in this cohort (kg C tree\(^{-1}\)).

seedC

Seed biomass of a tree in this cohort (kg C tree\(^{-1}\)).

leafC

Leaf biomass of a tree in this cohort (kg C tree\(^{-1}\)).

rootC

Fine root biomass of a tree in this cohort (kg C tree\(^{-1}\)).

SW_C

Sapwood biomass of a tree in this cohort (kg C tree\(^{-1}\)).

HW_C

Heartwood biomass of a tree in this cohort (kg C tree\(^{-1}\)).

NSN

Nonstructural nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

seedN

Seed nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

leafN

Leaf nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

rootN

Fine root nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

SW_N

Sapwood nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

HW_N

Heartwood nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

output_annual_tile

A data.frame with annual outputs at tile level.

year

Year of the simulation.

CAI

Crown area index (m\(^2\)/m\(^2\)).

LAI

Leaf area index (m\(^2\)/m\(^2\)).

Density

Number of trees per area (trees ha\(^{-1}\)).

DBH

Diameter at tile level (cm).

Density12

Tree density for trees with DBH > 12 cm (individuals ha\(^{-1}\)).

DBH12

Diameter at tile level considering trees with DBH > 12 cm (cm).

QMD

Quadratic mean diameter at tile level considering trees with DBH > 12 cm (cm).

NPP

Net primary productivity (kg C m\(^{-2}\) yr\(^{-1}\)).

GPP

Gross primary productivity (kg C m\(^{-2}\) yr\(^{-1}\)).

Rauto

Plant autotrophic respiration (kg C m\(^{-2}\) yr\(^{-1}\)).

Rh

Heterotrophic respiration (kg C m\(^{-2}\) yr\(^{-1}\)).

rain

Annual precipitation (mm m\(^{-2}\) yr\(^{-1}\)).

SoilWater

Soil water content in root zone (kg m\(^{-2}\)).

Transp

Transpiration (mm m\(^{-2}\) yr\(^{-1}\)).

Evap

Evaporation (mm m\(^{-2}\) yr\(^{-1}\)).

Runoff

Water runoff (mm m\(^{-2}\) yr\(^{-1}\)).

plantC

Plant biomass (kg C m\(^{-2}\)).

soilC

Soil carbon (kg C m\(^{-2}\)).

plantN

Plant nitrogen (kg N m\(^{-2}\)).

soilN

Soil nitrogen (kg N m\(^{-2}\)).

totN

Total nitrogen in plant and soil (kg N m\(^{-2}\)).

NSC

Nonstructural carbohydrates (kg C m\(^{-2}\)).

SeedC

Seed biomass (kg C m\(^{-2}\)).

leafC

Leaf biomass (kg C m\(^{-2}\)).

rootC

Fine root biomass (kg C m\(^{-2}\)).

SapwoodC

Sapwood biomass (kg C m\(^{-2}\)).

WoodC

Heartwood biomass (kg C m\(^{-2}\)).

NSN

Nonstructural nitrogen (kg N m\(^{-2}\)).

SeedN

Seed nitrogen (kg N m\(^{-2}\)).

leafN

Leaf nitrogen (kg N m\(^{-2}\)).

rootN

Fine root nitrogen (kg N m\(^{-2}\)).

SapwoodN

Sapwood nitrogen (kg N m\(^{-2}\)).

WoodN

Heartwood nitrogen (kg N m\(^{-2}\)).

McrbC

Microbial carbon (kg C m\(^{-2}\)).

fastSOM

Fast soil carbon pool (kg C m\(^{-2}\)).

SlowSOM

Slow soil carbon pool (kg C m\(^{-2}\)).

McrbN

Microbial nitrogen (kg N m\(^{-2}\)).

fastSoilN

Fast soil nitrogen pool (kg N m\(^{-2}\)).

slowsoilN

Slow soil nitrogen pool (kg N m\(^{-2}\)).

mineralN

Mineral nitrogen pool (kg N m\(^{-2}\)).

N_fxed

Nitrogen fixation (kg N m\(^{-2}\)).

N_uptk

Nitrogen uptake (kg N m\(^{-2}\)).

N_yrMin

Annual available nitrogen (kg N m\(^{-2}\)).

N_P25

Annual nitrogen from plants to soil (kg N m\(^{-2}\)).

N_loss

Annual nitrogen loss (kg N m\(^{-2}\)).

totseedC

Total seed carbon (kg C m\(^{-2}\)).

totseedN

Total seed nitrogen (kg N m\(^{-2}\)).

Seedling_C

Total carbon from all compartments but seeds (kg C m\(^{-2}\)).

Seeling_N

Total nitrogen from all compartments but seeds (kg N m\(^{-2}\)).

MaxAge

Age of the oldest tree in the tile (years).

MaxVolume

Maximum volumne of a tree in the tile (m\(^3\)).

MaxDBH

Maximum DBH of a tree in the tile (m).

NPPL

Growth of a tree, including carbon allocated to leaves (kg C m\(^{-2}\) year\(^{-1}\)).

NPPW

Growth of a tree, including carbon allocated to sapwood (kg C m\(^{-2}\) year\(^{-1}\)).

n_deadtrees

Number of trees that died (trees m\(^{-2}\) year\(^{-1}\)).

c_deadtrees

Carbon biomass of trees that died (kg C m\(^{-2}\) year\(^{-1}\)).

m_turnover

Continuous biomass turnover (kg C m\(^{-2}\) year\(^{-1}\)).

c_turnover_time

Carbon turnover rate, calculated as the ratio between plant biomass and NPP (year\(^{-1}\)).

output_annual_cohorts

A data.frame of annual outputs at the cohort level.

year

Year of the simulation.

cID

An integer indicating the cohort identity.

PFT

An integer indicating the Plant Functional Type.

layer

An integer indicating the crown layer, numbered from top to bottom.

density

Number of trees per area (trees ha\(^{-1}\)).

f_layer

Fraction of layer area occupied by this cohort.

dDBH

Diameter growth of a tree in this cohort (cm year\(^{-1}\)).

dbh

Tree diameter (cm).

height

Tree height (m).

age

Age of the cohort (years).

Acrow

Crown area of a tree in this cohort (m\(^2\)).

wood

Sum of sapwood and heartwood biomass of a tree in this cohort (kg C tree\(^{-1}\)).

nsc

Nonstructural carbohydrates in a tree (kg C tree\(^{-1}\)).

NSN

Nonstructural nitrogen of a tree (kg N tree\(^{-1}\)).

NPPtr

Total growth of a tree, including carbon allocated to seeds, leaves, fine roots, and sapwood (kg C tree\(^{-1}\) year\(^{-1}\)).

seed

Fraction of carbon allocated to seeds to total growth.

NPPL

Fraction of carbon allocated to leaves to total growth.

NPPR

Fraction of carbon allocated to fine roots to total growth.

NPPW

Fraction of carbon allocated to sapwood to total growth.

GPP_yr

Gross primary productivity of a tree (kg C tree\(^{-1}\) year\(^{-1}\)).

NPP_yr

Net primary productivity of a tree (kg C tree\(^{-1}\) year\(^{-1}\)).

Rauto

Plant autotrophic respiration (kg C tree\(^{-1}\) yr\(^{-1}\)).

N_uptk

Nitrogen uptake (kg N tree\(^{-1}\) yr\(^{-1}\)).

N_fix

Nitrogen fixation (kg N tree\(^{-1}\) yr\(^{-1}\)).

maxLAI

Maximum leaf area index for a tree (m\(^2\) m\(^{-2}\)).

Volume

Tree volume (m\(^3\)).

n_deadtrees

Number of trees that died (trees yr\(^{-1}\)).

c_deadtrees

Carbon biomass of trees that died (kg C yr\(^{-1}\)).

deathrate

Mortality rate of this cohort (yr\(^{-1}\)).

Examples

# \donttest{
# Example BiomeE model run

# Use example drivers data
drivers <- biomee_gs_leuning_drivers

# Run BiomeE for the first site
mod_output <- run_biomee_f_bysite(
 sitename = drivers$sitename[1],
 params_siml = drivers$params_siml[[1]],
 site_info = drivers$site_info[[1]],
 forcing = drivers$forcing[[1]],
 params_tile = drivers$params_tile[[1]],
 params_species = drivers$params_species[[1]],
 params_soil = drivers$params_soil[[1]],
 init_cohort = drivers$init_cohort[[1]],
 init_soil = drivers$init_soil[[1]]
)
# }