R wrapper for SOFUN

R wrapper for SOFUN

run_lm3ppa_f_bysite(
  sitename,
  params_siml,
  siteinfo,
  forcing,
  params_tile,
  params_species,
  params_soil,
  init_cohort,
  init_soil,
  makecheck = TRUE
)

Arguments

sitename
Site name

params_siml
Simulation parameters

siteinfo
Site meta info

forcing
forcing (input) dataframe (returned object by prepare_input_sofun)

params_tile
Tile-level model parameters

params_species
Species-specific model parameters

params_soil
Soil parameters (per soil layer)

init_cohort
Initial size of individuals in each of the N initally present cohorts

init_soil
Initial soil pool sizes

makecheck
A logical specifying whether checks are performed to verify forcings.

Details

This is the model