R wrapper for SOFUN P-model

Call to the Fortran P-model

run_pmodel_f_bysite(
  sitename,
  params_siml,
  site_info,
  forcing,
  params_soil,
  params_modl,
  makecheck = TRUE,
  verbose = TRUE
)

Arguments

sitename

Site name.

params_siml

Simulation parameters.

spinup

A logical value indicating whether this simulation does spin-up.

spinupyears

Number of spin-up years.

recycle

Length of standard recycling period, in days.

soilmstress

A logical value, if TRUE an empirical soil moisture stress function is applied to GPP.

tempstress

A logical value, if TRUE an empirical temperature stress function is applied to GPP.

calc_aet_fapar_vpd

(not in use)

in_ppfd

A logical value, if TRUE PPFD is a prescribed variable, if FALSE PPFD is simulated internally.

in_netrad

A logical value indicating whether net radiation is prescribed (TRUE) or simulated internally (FALSE).

outdt

An integer indicating the output periodicity.

ltre

A logical value, TRUE if evergreen tree.

ltne

A logical value, TRUE if evergreen tree and N-fixing.

ltrd

A logical value, TRUE if deciduous tree.

ltnd

A logical value, TRUE if deciduous tree and N-fixing.

lgr3

A logical value, TRUE if grass with C3 photosynthetic pathway.

lgn3

A logical value, TRUE if grass with C3 photosynthetic pathway and N-fixing.

lgr4

A logical value, TRUE if grass with C4 photosynthetic pathway.

firstyeartrend

The year AD of first transient year.

nyeartrend

The number of transient years.

site_info

A list of site meta info. Required:

lon

Longitud of the site location.

lat

Latitude of the site location.

elv

Elevation of the site location, in meters.

whc

A numeric value for the water holding capacity (in mm), used for simulating the soil water balance.

forcing

A data frame of forcing climate data, used as input (returned object by collect_drivers_sofun, see p_model_drivers for a detailed description of its structure and contents).

params_soil

A list of soil texture parameters, for the top and bottom layer of soil.

fsand

The fraction of sand in the soil.

fclay

The fraction of clay in the soil.

forg

The fraction of organic matter in the soil.

fgravel

The fraction of gravel in the soil.

params_modl

A named list of free (calibratable) model parameters.

kphio

The quantum yield efficiency parameter, in mol mol\(^{-1}\).

soilm_par_a

Intercept of the linear soil moisture stress function (unitless).

soilm_par_b

Slope of the linear soil moisture stress function (unitless).

tau_acclim_tempstress

Memory e-folding time scale parameter (in hours) for acclimation of photosynthesis, following Eq. 5 in Makela et al. (2004) Tree Physiology 24, 369–376

par_shape_tempstress

Shape parameter for the low temperature stress function affecting quantum yield efficiency. A positive (unitless) value indicating the sensitivity of the decline (assuming no decline at 10\(^o\)C) and above, but decreasing with a quadratic function below 10\(^o\)C).

makecheck

A logical specifying whether checks are performed to verify forcings and model parameters. TRUE by default.

verbose

A logical specifying whether to print warnings. Defaults to TRUE.

Value

Model output is provided as a tidy dataframe, with columns:

date

Date of the observation in YYYY-MM-DD format.

year_dec

Decimal representation of year and day of the year (for example, 2007.000 corresponds to 2007-01-01 and 2007.003 to 2007-01-02.

fapar

Fraction of photosynthetic active radiation (fAPAR), taking values between 0 and 1.

gpp

Gross Primary Productivity (GPP) for each time stamp (in gC m\(^{-2}\) d\(^{-1}\)).

transp

Actual evapotranspiration (AET) (in mm d\(^{-1}\)).

latenth

Latent heat flux (in J m\(^{-2}\) d\(^{-1}\)).

pet

Potential evapotranspiration (PET) (in mm d\(^{-1}\)).

vcmax

Maximum rate of RuBisCO carboxylation (Vcmax) (in mol C m\(^{-2}\) d\(^{-1}\)).

jmax

Maximum rate of electron transport for RuBP regeneration (in mol CO\(_2\) m\(^{-2}\) s\(^{-1}\)).

vcmax25

Maximum rate of carboxylation (Vcmax), normalised to 25\(^o\)C (in mol C m\(^{-2}\) d\(^{-1}\)).

jmax25

Maximum rate of electron transport, normalised to 25\(^o\)C (in mol C m\(^{-2}\) s\(^{-1}\)).

gs_accl

Acclimated stomatal conductance (in mol C m\(^{−2}\) d\(^{−1}\) Pa\(^{−1}\)).

wscal

Relative soil water content, between 0 (permanent wilting point, PWP) and 1 (field capacity, FC).

chi

Ratio of leaf-internal to ambient CO\(_{2}\), ci:ca (unitless).

iwue

Intrinsic water use efficiency (iWUE) (in Pa).

rd

Dark respiration (Rd) in gC m\(^{-2}\) d\(^{-1}\).

tsoil

Soil temperature, in \(^{o}\)C.