Run LM3-PPA

Runs the LM3-PPA model and loads output in once.

runread_lm3ppa_f(
  df_drivers,
  params_modl,
  makecheck = TRUE,
  parallel = FALSE,
  ncores = 2
)

Arguments

df_drivers
A nested data frame with one row for each site and columns named according to the arguments of function runread_pmodel_f_bysite

params_modl
A named list of model parameters

makecheck
A logical specifying whether checks are performed to verify forcings.

parallel
A logical specifying whether simulations are to be parallelised (sending data from a certain number of sites to each core). Defaults to FALSE.

ncores
An integer specifying the number of cores used for parallel computing. Defaults to 2.

Value

A data frame (tibble) with one row for each site and outputs stored in the nested column data

Examples

mod <- runread_pmodel_f( df_drivers, params_modl, makecheck = TRUE, parallel = FALSE, ncores = 2 )

#> Error: Problem with `mutate()` input `data`.
#> x Can't subset columns that don't exist.
#> x Column `rain` doesn't exist.
#> ℹ Input `data` is `purrr::pmap(...)`.