rsofun BiomeE driver data (Leuning photosynthesis model)

Small dataset representing the driver to run the BiomeE-model at the CH-LAE site using the Leuning photosynthesis specification (and half-hourly time step) It can also be used together with leaf trait data from CH-LAE (biomee_validation) to optimize model parameters.

biomee_gs_leuning_drivers

Format

A tibble of driver data.

sitename

Site name

params_siml

Simulation parameters as a data.frame, including the following data:

spinup

Flag indicating whether this simulation does spin-up.

spinupyears

Number of spin-up years.

recycle

Number of first N years of forcing data.frame that are recycled for spin-up.

firstyeartrend

Year of first transient year (AD) (optional). Is only used to set years in output data frames. Defaults to 0 if not provided.

nyeartrend

Number of transient years (optional). Determines the length of simulation output after spin-up. Defaults to number of years contained in the forcing data. (If longer than forcing data, last year of forcing is repeated until the end (spin-down).)

steps_per_day

Time resolution of the forcing (day-1).

do_U_shaped_mortality

Flag indicating whether U-shaped mortality is used.

update_annualLAImax

Flag indicating whether updating LAImax according to mineral N in soil.

do_closedN_run

Flag indicating whether doing N closed runs to recover N balance enforcing 0.2 kg N m-2 in the inorganic N pool.

code_method_photosynth

String specifying the method of photosynthesis used in the model, either "pmodel" or "gs_leuning".document()

code_method_mortality

String indicating the type of mortality in the model. One of the following: "dbh" is size-dependent mortality, "const_selfthin" is constant self thinning (in development), "cstarvation" is carbon starvation, and "growthrate" is growth rate dependent mortality.

site_info

Site meta info in a data.frame. This data structure can be freely used for documenting the dataset, but must include at least the following data:

lon

Longitude of the site location in degrees east.

lat

Latitude of the site location in degrees north.

elv

Elevation of the site location, in meters above sea level.

forcing

Forcing data.frame used as input

ppfd

Photosynthetic photon flux density (mol s-1 m-2)

tair

Air temperature (deg C)

vpd

Vapor pressure deficit (Pa)

rain

Precipitation (kgH2O m-2 s-1 == mm s-1)

wind

Wind velocity (m s-1)

pair

Atmospheric pressure (pa)

co2

CO2 atmospheric concentration (ppm)

params_tile

Tile-level model parameters, into a single row data.frame, including the following data:

soiltype

Integer indicating the type of soil: Sand = 1, LoamySand = 2, SandyLoam = 3, SiltLoam = 4, FrittedClay = 5, Loam = 6, Clay = 7.

FLDCAP

Field capacity (vol/vol). Water remaining in a soil after it has been thoroughly saturated and allowed to drain freely.

WILTPT

Wilting point (vol/vol). Water content of a soil at which plants wilt and fail to recover.

K1

Fast soil C decomposition rate (year\(^{-1}\)).

K2

Slow soil C decomposition rate (year\(^{-1}\)).

K_nitrogen

Mineral Nitrogen turnover rate (year\(^{-1}\)).

MLmixRatio

Ratio of C and N returned to litters from microbes.

etaN

N loss rate through runoff (organic and mineral) (year\(^{-1}\)).

LMAmin

Minimum LMA, leaf mass per unit area, kg C m\(^{-2}\).

fsc_fine

Fraction of fast turnover carbon in fine biomass.

fsc_wood

Fraction of fast turnover carbon in wood biomass.

GR_factor

Growth respiration factor.

l_fract

Fraction of the carbon retained after leaf drop.

retransN

Retranslocation coefficient of nitrogen.

f_initialBSW

Coefficient for setting up initial sapwood.

f_N_add

Re-fill of N for sapwood.

tf_base

Calibratable scalar for respiration, used to increase LUE levels.

par_mort

Canopy mortality parameter.

par_mort_under

Parameter for understory mortality.

params_species

A data.frame containing species-specific model parameters, with one species per row, including the following data:

lifeform

Integer set to 0 for grasses and 1 for trees.

phenotype

Integer set to 0 for deciduous and 1 for evergreen.

pt

Integer indicating the type of plant according to photosynthesis: 0 for C3; 1 for C4

alpha_FR

Fine root turnover rate (year\(^{-1}\)).

rho_FR

Material density of fine roots (kg C m\(^{-3}\)).

root_r

Radius of the fine roots, in m.

root_zeta

e-folding parameter of root vertical distribution, in m.

Kw_root

Fine root water conductivity (mol m\(^{-2}\) s\(^{-1}\) MPa\(^{-1}\)).

leaf_size

Characteristic leaf size.

Vmax

Max RuBisCo rate, in mol m\(^{-2}\) s\(^{-1}\).

Vannual

Annual productivity per unit area at full sun (kg C m\(^{-2}\) year\(^{-2}\)).

wet_leaf_dreg

Wet leaf photosynthesis down-regulation.

m_cond

Factor of stomatal conductance.

alpha_phot

Photosynthesis efficiency.

gamma_L

Leaf respiration coefficient, in year\(^{-1}\).

gamma_LN

Leaf respiration coefficient per unit N.

gamma_SW

Sapwood respiration rate, in kg C m\(^{-2}\) year\(^{-1}\).

gamma_FR

Fine root respiration rate, kg C kg C\(^{-1}\) year\(^{-1}\).

tc_crit

Critical temperature triggerng offset of phenology, in Kelvin.

tc_crit_on

Critical temperature triggerng onset of phenology, in Kelvin.

gdd_crit

Critical value of GDD5 for turning ON growth season.

betaON

Critical soil moisture for phenology onset.

betaOFF

Critical soil moisture for phenology offset.

seedlingsize

Initial size of seedlings, in kg C per individual.

LNbase

Basal leaf N per unit area, in kg N m\(^{-2}\).

lAImax

Maximum crown LAI (leaf area index).

Nfixrate0

Reference N fixation rate (kg N kg C\(^{-1}\) root).

NfixCost0

Carbon cost of N fixation (kg C kg N\(^{-1}\)).

phiCSA

Ratio of sapwood area to leaf area.

mortrate_d_c

Canopy tree mortality rate (year\(^{-1}\)).

mortrate_d_u

Understory tree mortality rate (year\(^{-1}\)).

maturalage

Age at which trees can reproduce (years).

v_seed

Fraction of G_SF to G_F.

fNSmax

Multiplier for NSNmax as sum of potential bl and br.

LMA

Leaf mass per unit area (kg C m\(^{-2}\)).

rho_wood

Wood density (kg C m\(^{-3}\)).

alphaBM

Coefficient for allometry (biomass = alphaBM * DBH ** thetaBM).

thetaBM

Coefficient for allometry (biomass = alphaBM * DBH ** thetaBM).

kphio

Quantum yield efficiency \(\varphi_0\), in mol mol\(^{-1}\).

phiRL

Ratio of fine root to leaf area.

LAI_light

Maximum LAI limited by light.

init_cohort

A data.frame of initial cohort specifications, including the following data:

init_cohort_species

Index of a species described in param_species.

init_cohort_nindivs

Initial individual density, in individuals per m\(^{2}\).

init_cohort_bl

Initial biomass of leaf, in kg C per individual.

init_cohort_br

Initial biomass of fine root, in kg C per individual.

init_cohort_bsw

Initial biomass of sapwood, in kg C per individual.

init_cohort_bHW

Initial biomass of heartwood, in kg C per individual.

init_cohort_seedC

Initial biomass of seed, in kg C per individual.

init_cohort_nsc

Initial non-structural biomass, in kg C per individual.

init_soil

A data.frame of initial soil pools, including the following data:

init_fast_soil_C

Initial fast soil carbon, in kg C m\(^{-2}\).

init_slow_soil_C

Initial slow soil carbon, in kg C m\(^{-2}\).

init_Nmineral

Mineral nitrogen pool, in kg N m\(^{-2}\).

N_input

Annual nitrogen input to soil N pool, in kg N m\(^{-2}\) year\(^{-1}\).