rsofun p-model driver data (for leaf traits)

Small tests dataset to validate if compiled code and optimization routines can run for leaf traits data

p_model_drivers_vcmax25

Format

A tibble of model driver data:

sitename

A character string containing the site names.

forcing

A tibble of forcing climate data for an "average" year, including the following variables:

date

Date in YYYY-MM-DD format (here representative of the day of the year instead of the date of observation).

temp

Air temperature in \(^\circ\)C.

vpd

Vapour pressure deficit in Pa.

ppfd

Photosynthetic photon flux density (PPFD) in mol m\(^{-2}\) s\(^{-1}\). If all values are NA, it indicates that PPFD should be calculated by the SPLASH model.

netrad

Net radiation in W m\(^{-2}\). This is currently ignored as a model forcing.

patm

Atmospheric pressure in Pa.

ccov

Cloud coverage in %. This is only used when either PPFD or net radiation are not prescribed.

snow

Snow in mm d\(^{-1}\).

rain

Rain in mm d\(^{-1}\).

fapar

Fraction of photosynthetic active radiation (fAPAR), taking values between 0 and 1.

co2

Annually varying observed atmospheric CO\(_2\), identical across sites.

tmin

Daily minimum air temperature in \(^\circ\)C (set equal to temp).

tmax

Daily maximum air temperature in \(^\circ\)C.(set equal to temp).

params_siml

A tibble containing simulation parameters.

spinup

A logical value indicating whether this simulation does spin-up.

spinupyears

Number of spin-up years.

recycle

Length of standard recycling period, in days.

outdt

An integer indicating the output periodicity.

ltre

A logical value, TRUE if evergreen tree.

ltne

A logical value, TRUE if evergreen tree and N-fixing.

ltrd

A logical value, TRUE if deciduous tree.

ltnd

A logical value, TRUE if deciduous tree and N-fixing.

lgr3

A logical value, TRUE if grass with C3 photosynthetic pathway.

lgn3

A logical value, TRUE if grass with C3 photosynthetic pathway and N-fixing.

lgr4

A logical value, TRUE if grass with C4 photosynthetic pathway.

site_info

A tibble containing site meta information.

lon

Longitud of the site location.

lat

Latitude of the site location.

elv

Elevation of the site location, in meters.

whc

A numeric value for the root zone water holding capacity (in mm), used for simulating the soil water balance.

Source

Atkin, O. K., Bloomfield, K. J., Reich, P. B., Tjoelker, M. G., Asner, G. P., Bonal, D., et al. (2015). Global variability in leaf respiration in relation to climate, plant functional types and leaf traits. New Phytol. 206 (2), 614–636. doi:10.1111/nph.13253

University of East Anglia Climatic Research Unit; Harris, I.C.; Jones, P.D.; Osborn, T. (2021): CRU TS4.05: Climatic Research Unit (CRU) Time-Series (TS) version 4.05 of high-resolution gridded data of month-by-month variation in climate (Jan. 1901- Dec. 2020). NERC EDS Centre for Environmental Data Analysis, date of citation. https://catalogue.ceda.ac.uk/uuid/c26a65020a5e4b80b20018f148556681

Weedon, G. P., G. Balsamo, N. Bellouin,S. Gomes, M. J. Best, and P. Viterbo(2014), The WFDEI meteorologicalforcing data set: WATCH Forcing Datamethodology applied to ERA-Interimreanalysis data, Water Resour. Res.,50,7505–7514, doi:10.1002/2014WR015638.

Fick, S.E. and R.J. Hijmans, 2017. WorldClim 2: new 1km spatial resolution climate surfaces for global land areas. International Journal of Climatology 37 (12): 4302-4315.

C.D. Keeling, R.B. Bacastow, A.E. Bainbridge, C.A. Ekdahl, P.R. Guenther, and L.S. Waterman, (1976), Atmospheric carbon dioxide variations at Mauna Loa Observatory, Hawaii, Tellus, vol. 28, 538-551