Run BiomeE (R wrapper)

Run BiomeE Fortran model on single site.

run_biomee_f_bysite(
  sitename,
  params_siml,
  site_info,
  forcing,
  params_tile,
  params_species,
  init_cohort,
  init_soil,
  makecheck = TRUE
)

Arguments

sitename

Site name.

params_siml

Simulation parameters.

site_info

Site meta info in a data.frame.

forcing

A data frame of forcing climate data, used as input.

params_tile

Tile-level model parameters, into a single row data.frame.

params_species

A data.frame containing species-specific model parameters, with one species per row. See examples biomee_gs_leuning_drivers or biomee_p_model_drivers

init_cohort

A data.frame of initial cohort specifications.

init_soil

A data.frame of initial soil pools.

makecheck

A logical specifying whether checks are performed to verify forcings and model parameters. TRUE by default.

For further specifications of above inputs and examples see biomee_gs_leuning_drivers or biomee_p_model_drivers

Value

Model output is provided as a list, with elements:

output_hourly_tile

A data.frame containing hourly predictions .

year

Year of the simulation.

doy

Day of the year.

hour

Hour of the day.

rad

Radiation, in W m\(^{-2}\).

Tair

Air temperature, in Kelvin.

Prcp

Precipitation, in mm m\(^{-2}\).

GPP

Gross primary production (kg C m\(^{-2}\) hour\(^{-1}\)).

Resp

Plant respiration (kg C m\(^{-2}\) hour\(^{-1}\)).

Transp

Transpiration (mm m\(^{-2}\)).

Evap

Evaporation (mm m\(^{-2}\)).

Runoff

Water runoff (mm m\(^{-2}\)).

Soilwater

Soil water content in root zone (kg m\(^{-2}\)).

wcl

Volumetric soil water content for each layer (vol/vol).

FLDCAP

Field capacity (vol/vol).

WILTPT

Wilting point (vol/vol).

output_daily_tile

A data.frame with daily outputs at a tile level.

year

Year of the simulation.

doy

Day of the year.

Tc

Air temperature (Kelvin).

Prcp

Precipitation (mm m\(^{-2}\)).

totWs

Soil water content in root zone (kg m\(^{-2}\)).

Trsp

Transpiration (mm m\(^{2-}\)).

Evap

Evaporation (mm m\(^{-2}\)).

Runoff

Water runoff (mm m\(^{-2}\)).

ws1

Volumetric soil water content for layer 1.

ws2

Volumetric soil water content for layer 2.

ws3

Volumetric soil water content for layer 3.

LAI

Leaf area index (m\(^2\)/m\(^2\)).

GPP

Gross primary production (kg C m\(^{-2}\) day\(^{-1}\)).

Rauto

Plant autotrophic respiration (kg C m\(^{-2}\) day\(^{-1}\)).

Rh

Heterotrophic respiration (kg C m\(^{-2}\) day\(^{-1}\)).

NSC

Non-structural carbon (kg C m\(^{-2}\)).

seedC

Biomass of seeds (kg C m\(^{-2}\)).

leafC

Biomass of leaves (kg C m\(^{-2}\)).

rootC

Biomass of fine roots (kg C m\(^{-2}\)).

SW_C

Biomass of sapwood (kg C m\(^{-2}\)).

HW_C

biomass of heartwood (kg C m\(^{-2}\)).

NSN

Non-structural N pool (kg N m\(^{-2}\)).

seedN

Nitrogen of seeds (kg N m\(^{-2}\)).

leafN

Nitrogen of leaves (kg N m\(^{-2}\)).

rootN

Nitrogen of roots (kg N m\(^{-2}\)).

SW_N

Nitrogen of sapwood (kg N m\(^{-2}\)).

HW_N

Nitrogen of heartwood (kg N m\(^{-2}\)).

McrbC

Microbial carbon (kg C m\(^{-2}\)).

fastSOM

Fast soil carbon pool (kg C m\(^{-2}\)).

slowSOM

Slow soil carbon pool (kg C m\(^{-2}\)).

McrbN

Microbial nitrogen (kg N m\(^{-2}\)).

fastSoilN

Fast soil nitrogen pool (kg N m\(^{-2}\)).

slowSoilN

Slow soil nitrogen pool (kg N m\(^{-2}\)).

mineralN

Mineral nitrogen pool (kg N m\(^{-2}\)).

N_uptk

Nitrogen uptake (kg N m\(^{-2}\)).

output_daily_cohorts

A data.frame with daily predictions for each canopy cohort.

year

Year of the simulation.

doy

Day of the year.

hour

Hour of the day.

cID

An integer indicating the cohort identity.

PFT

An integer indicating the Plant Functional Type.

layer

An integer indicating the crown layer, numbered from top to bottom.

density

Number of trees per area (trees ha\(^{-1}\)).

f_layer

Fraction of layer area occupied by this cohort.

LAI

Leaf area index (m\(^2\)/m\(^2\)).

gpp

Gross primary productivity (kg C tree\(^{-1}\) day\(^{-1}\)).

resp

Plant autotrophic respiration (kg C tree\(^{-1}\) day\(^{-1}\)).

transp

Transpiration (mm tree\(^{-1}\) day\(^{-1}\)).

NPPleaf

Carbon allocated to leaves (kg C tree\(^{-1}\) day\(^{-1}\)).

NPProot

Carbon allocated to fine roots (kg C tree\(^{-1}\) day\(^{-1}\)).

NPPwood

Carbon allocated to wood (kg C tree\(^{-1}\) day\(^{-1}\)).

NSC

Nonstructural carbohydrates of a tree in this cohort (kg C tree\(^{-1}\)).

seedC

Seed biomass of a tree in this cohort (kg C tree\(^{-1}\)).

leafC

Leaf biomass of a tree in this cohort (kg C tree\(^{-1}\)).

rootC

Fine root biomass of a tree in this cohort (kg C tree\(^{-1}\)).

SW_C

Sapwood biomass of a tree in this cohort (kg C tree\(^{-1}\)).

HW_C

Heartwood biomass of a tree in this cohort (kg C tree\(^{-1}\)).

NSN

Nonstructural nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

seedN

Seed nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

leafN

Leaf nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

rootN

Fine root nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

SW_N

Sapwood nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

HW_N

Heartwood nitrogen of a tree in this cohort (kg N tree\(^{-1}\)).

output_annual_tile

A data.frame with annual outputs at tile level.

year

Year of the simulation.

CAI

Crown area index (m\(^2\)/m\(^2\)).

LAI

Leaf area index (m\(^2\)/m\(^2\)).

Density

Number of trees per area (trees ha\(^{-1}\)).

DBH

Diameter at tile level (cm).

Density12

Tree density for trees with DBH > 12 cm (individuals ha\(^{-1}\)).

DBH12

Diameter at tile level considering trees with DBH > 12 cm (cm).

QMD12

Quadratic mean diameter at tile level considering trees with DBH > 12 cm (cm).

NPP

Net primary productivity (kg C m\(^{-2}\) yr\(^{-1}\)).

GPP

Gross primary productivity (kg C m\(^{-2}\) yr\(^{-1}\)).

Rauto

Plant autotrophic respiration (kg C m\(^{-2}\) yr\(^{-1}\)).

Rh

Heterotrophic respiration (kg C m\(^{-2}\) yr\(^{-1}\)).

rain

Annual precipitation (mm m\(^{-2}\) yr\(^{-1}\)).

SoilWater

Soil water content in root zone (kg m\(^{-2}\)).

Transp

Transpiration (mm m\(^{-2}\) yr\(^{-1}\)).

Evap

Evaporation (mm m\(^{-2}\) yr\(^{-1}\)).

Runoff

Water runoff (mm m\(^{-2}\) yr\(^{-1}\)).

plantC

Plant biomass (kg C m\(^{-2}\)).

soilC

Soil carbon (kg C m\(^{-2}\)).

plantN

Plant nitrogen (kg N m\(^{-2}\)).

soilN

Soil nitrogen (kg N m\(^{-2}\)).

totN

Total nitrogen in plant and soil (kg N m\(^{-2}\)).

NSC

Nonstructural carbohydrates (kg C m\(^{-2}\)).

SeedC

Seed biomass (kg C m\(^{-2}\)).

leafC

Leaf biomass (kg C m\(^{-2}\)).

rootC

Fine root biomass (kg C m\(^{-2}\)).

SapwoodC

Sapwood biomass (kg C m\(^{-2}\)).

WoodC

Heartwood biomass (kg C m\(^{-2}\)).

NSN

Nonstructural nitrogen (kg N m\(^{-2}\)).

SeedN

Seed nitrogen (kg N m\(^{-2}\)).

leafN

Leaf nitrogen (kg N m\(^{-2}\)).

rootN

Fine root nitrogen (kg N m\(^{-2}\)).

SapwoodN

Sapwood nitrogen (kg N m\(^{-2}\)).

WoodN

Heartwood nitrogen (kg N m\(^{-2}\)).

McrbC

Microbial carbon (kg C m\(^{-2}\)).

fastSOM

Fast soil carbon pool (kg C m\(^{-2}\)).

SlowSOM

Slow soil carbon pool (kg C m\(^{-2}\)).

McrbN

Microbial nitrogen (kg N m\(^{-2}\)).

fastSoilN

Fast soil nitrogen pool (kg N m\(^{-2}\)).

slowsoilN

Slow soil nitrogen pool (kg N m\(^{-2}\)).

mineralN

Mineral nitrogen pool (kg N m\(^{-2}\)).

N_fxed

Nitrogen fixation (kg N m\(^{-2}\)).

N_uptk

Nitrogen uptake (kg N m\(^{-2}\)).

N_yrMin

Annual available nitrogen (kg N m\(^{-2}\)).

N_P25

Annual nitrogen from plants to soil (kg N m\(^{-2}\)).

N_loss

Annual nitrogen loss (kg N m\(^{-2}\)).

totseedC

Total seed carbon (kg C m\(^{-2}\)).

totseedN

Total seed nitrogen (kg N m\(^{-2}\)).

Seedling_C

Total carbon from all compartments but seeds (kg C m\(^{-2}\)).

Seeling_N

Total nitrogen from all compartments but seeds (kg N m\(^{-2}\)).

MaxAge

Age of the oldest tree in the tile (years).

MaxVolume

Maximum volumne of a tree in the tile (m\(^3\)).

MaxDBH

Maximum DBH of a tree in the tile (m).

NPPL

Growth of a tree, including carbon allocated to leaves (kg C m\(^{-2}\) year\(^{-1}\)).

NPPW

Growth of a tree, including carbon allocated to sapwood (kg C m\(^{-2}\) year\(^{-1}\)).

n_deadtrees

Number of trees that died (trees m\(^{-2}\) year\(^{-1}\)).

c_deadtrees

Carbon biomass of trees that died (kg C m\(^{-2}\) year\(^{-1}\)).

m_turnover

Continuous biomass turnover (kg C m\(^{-2}\) year\(^{-1}\)).

c_turnover_time

Carbon turnover rate, calculated as the ratio between plant biomass and NPP (year\(^{-1}\)).

output_annual_cohorts

A data.frame of annual outputs at the cohort level.

year

Year of the simulation.

cID

An integer indicating the cohort identity.

PFT

An integer indicating the Plant Functional Type.

layer

An integer indicating the crown layer, numbered from top to bottom.

density

Number of trees per area (trees ha\(^{-1}\)).

f_layer

Fraction of layer area occupied by this cohort.

dDBH

Diameter growth of a tree in this cohort (cm year\(^{-1}\)).

dbh

Tree diameter (cm).

height

Tree height (m).

age

Age of the cohort (years).

Acrow

Crown area of a tree in this cohort (m\(^2\)).

wood

Sum of sapwood and heartwood biomass of a tree in this cohort (kg C tree\(^{-1}\)).

nsc

Nonstructural carbohydrates in a tree (kg C tree\(^{-1}\)).

NSN

Nonstructural nitrogen of a tree (kg N tree\(^{-1}\)).

NPPtr

Total growth of a tree, including carbon allocated to seeds, leaves, fine roots, and sapwood (kg C tree\(^{-1}\) year\(^{-1}\)).

seed

Fraction of carbon allocated to seeds to total growth.

NPPL

Fraction of carbon allocated to leaves to total growth.

NPPR

Fraction of carbon allocated to fine roots to total growth.

NPPW

Fraction of carbon allocated to sapwood to total growth.

GPP_yr

Gross primary productivity of a tree (kg C tree\(^{-1}\) year\(^{-1}\)).

NPP_yr

Net primary productivity of a tree (kg C tree\(^{-1}\) year\(^{-1}\)).

Rauto

Plant autotrophic respiration (kg C tree\(^{-1}\) yr\(^{-1}\)).

N_uptk

Nitrogen uptake (kg N tree\(^{-1}\) yr\(^{-1}\)).

N_fix

Nitrogen fixation (kg N tree\(^{-1}\) yr\(^{-1}\)).

maxLAI

Maximum leaf area index for a tree (m\(^2\) m\(^{-2}\)).

Volume

Tree volume (m\(^3\)).

n_deadtrees

Number of trees that died (trees yr\(^{-1}\)).

c_deadtrees

Carbon biomass of trees that died (kg C yr\(^{-1}\)).

deathrate

Mortality rate of this cohort (yr\(^{-1}\)).

Examples

# \donttest{
# Example BiomeE model run

# Use example drivers data
drivers <- biomee_gs_leuning_drivers

# Run BiomeE for the first site
mod_output <- run_biomee_f_bysite(
 sitename = drivers$sitename[1],
 params_siml = drivers$params_siml[[1]],
 site_info = drivers$site_info[[1]],
 forcing = drivers$forcing[[1]],
 params_tile = drivers$params_tile[[1]],
 params_species = drivers$params_species[[1]],
 init_cohort = drivers$init_cohort[[1]],
 init_soil = drivers$init_soil[[1]]
)
# }