R wrapper for SOFUN P-model

Call to the Fortran P-model

run_pmodel_f_bysite(
  sitename,
  params_siml,
  site_info,
  forcing,
  params_soil,
  params_modl,
  makecheck = TRUE,
  verbose = TRUE
)

Arguments

sitename

Site name.

params_siml

Simulation parameters.

spinup

A logical value indicating whether this simulation does spin-up.

spinupyears

Number of spin-up years.

recycle

Length of standard recycling period, in days.

soilmstress

A logical value, if TRUE an empirical soil moisture stress function is applied to GPP.

tempstress

A logical value, if TRUE an empirical temperature stress function is applied to GPP.

calc_aet_fapar_vpd

(not in use)

in_ppfd

A logical value, if TRUE PPFD is a prescribed variable, if FALSE PPFD is simulated internally.

in_netrad

A logical value indicating whether net radiation is prescribed (TRUE) or simulated internally (FALSE).

outdt

An integer indicating the output periodicity.

ltre

A logical value, TRUE if evergreen tree.

ltne

A logical value, TRUE if evergreen tree and N-fixing.

ltrd

A logical value, TRUE if deciduous tree.

ltnd

A logical value, TRUE if deciduous tree and N-fixing.

lgr3

A logical value, TRUE if grass with C3 photosynthetic pathway.

lgn3

A logical value, TRUE if grass with C3 photosynthetic pathway and N-fixing.

lgr4

A logical value, TRUE if grass with C4 photosynthetic pathway.

firstyeartrend

The year AD of first transient year.

nyeartrend

The number of transient years.

site_info

A list of site meta info. Required:

lon

Longitud of the site location.

lat

Latitude of the site location.

elv

Elevation of the site location, in meters.

whc

A numeric value for the water holding capacity (in mm), used for simulating the soil water balance.

forcing

A data frame of forcing climate data, used as input (returned object by collect_drivers_sofun).

params_soil

A list of soil texture parameters, for the top and bottom layer of soil.

fsand

The fraction of sand in the soil.

fclay

The fraction of clay in the soil.

forg

The fraction of organic matter in the soil.

fgravel

The fraction of gravel in the soil.

params_modl

A named list of free (calibratable) model parameters.

kphio

The quantum yield efficiency parameter, in mol mol\(^{-1}\).

soilm_par_a

Intercept of the linear soil moisture stress function (unitless).

soilm_par_b

Slope of the linear soil moisture stress function (unitless).

tau_acclim_tempstress

Memory e-folding time scale parameter (in hours) for acclimation of photosynthesis, following Eq. 5 in Makela et al. (2004) Tree Physiology 24, 369–376

par_shape_tempstress

Shape parameter for the low temperature stress function affecting quantum yield efficiency. A positive (unitless) value indicating the sensitivity of the decline (assuming no decline at 10\(^o\)C) and above, but decreasing with a quadratic function below 10\(^o\)C).

makecheck

A logical specifying whether checks are performed to verify forcings and model parameters. TRUE by default.

verbose

A logical specifying whether to print warnings. Defaults to TRUE.

Details

Model output is provided as a tidy dataframe, with columns "fapar", "gpp", "transp", "latenth", "pet", "vcmax", "jmax", "vcmax25", "jmax25", "gs_accl", "wscal", "chi", "iwue", "rd".